refactor: extra app entry
