+++ src/App.tsx
@@ -1,17 +1,22 @@
-import { Canvas } from './component/canvas';
-import { Designer } from './component/designer';
-import { Toolsbar } from './component/toolsbar';
+import { NextUIProvider } from '@nextui-org/react';
+import { Provider } from 'jotai';
+import { DevTools } from 'jotai-devtools';
+import { BrowserRouter, Route, Routes } from 'react-router-dom';
+import { Preview } from './page/preview';
+import { appStore } from './store';
+import './styles.css';
+import { Main } from './page/main';
 
-function App() {
-  return (
-    <div className='flex h-full'>
-      <div className='h-full flex-1 border-r-1 border-solid border-[#ddd]'>
-        <Toolsbar />
-        <Canvas />
-      </div>
-      <Designer />
-    </div>
-  );
-}
-
-export default App;
+export const App = () => (
+  <BrowserRouter>
+    <NextUIProvider className='h-full'>
+      <Provider store={appStore}>
+        <Routes>
+          <Route path='/' element={<Main />} />
+          <Route path='/preview' element={<Preview />} />
+        </Routes>
+      </Provider>
+      <DevTools store={appStore} />
+    </NextUIProvider>
+  </BrowserRouter>
+);
