+++ src/page/main.tsx
@@ -1,0 +1,15 @@
+import { Canvas } from '../component/canvas';
+import { Designer } from '../component/designer';
+import { Toolsbar } from '../component/toolsbar';
+
+export const Main = () => {
+  return (
+    <div className='flex h-full'>
+      <div className='h-full flex-1 border-r-1 border-solid border-[#ddd]'>
+        <Toolsbar />
+        <Canvas />
+      </div>
+      <Designer />
+    </div>
+  );
+};
